python/requirements.txt
django
djangorestframework
django-rest-auth
django-allauth
